Commit and push July 18_1st
