+++ New BDD Project/src/main/java/com/QM/utils/DecryptionUtil.java
@@ -1,0 +1,8 @@
+package com.QM.utils;
+
+
+public class DecryptionUtil {
+    public static String decrypt(String encryptedPassword) throws Exception {
+        return DataDecrypt.decrypt(encryptedPassword);
+    }
+}
