+++ New BDD Project/src/main/java/com/QM/utils/ExcelUtil.java
@@ -1,0 +1,49 @@
+package com.QM.utils;
+
+import org.apache.poi.ss.usermodel.*;
+import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import java.io.FileInputStream;
+import java.io.IOException;
+
+public class ExcelUtil {
+    private Workbook workbook;
+    private Sheet sheet;
+
+    public ExcelUtil(String filePath, String sheetName) throws IOException {
+        FileInputStream file = new FileInputStream(filePath);
+        workbook = new XSSFWorkbook(file);
+        sheet = workbook.getSheet(sheetName);
+    }
+
+    public String getCellData(int rowNum, int colNum) {
+        Row row = sheet.getRow(rowNum);
+        Cell cell = row.getCell(colNum);
+
+        switch (cell.getCellType()) {
+            case CellType.STRING:
+                return cell.getStringCellValue();
+            case CellType.NUMERIC:
+                if (DateUtil.isCellDateFormatted(cell)) {
+                    return cell.getDateCellValue().toString();
+                } else {
+                    return String.valueOf(cell.getNumericCellValue());
+                }
+            case CellType.BOOLEAN:
+                return String.valueOf(cell.getBooleanCellValue());
+            case CellType.FORMULA:
+                return cell.getCellFormula();
+            default:
+                return "";
+        }
+    }
+
+    public int getRowCount() {
+        return sheet.getLastRowNum() + 1;
+    }
+
+    public int getColumnCount() {
+        return sheet.getRow(0).getLastCellNum();
+    }
+}
+
+
