--- New BDD Project/src/test/java/pages/LoginPage.java
@@ -1,49 +1,0 @@
-package pages;
-
-import org.openqa.selenium.Keys;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
-
-public class LoginPage {
-WebDriver driver;
-
-    @FindBy(id = "username")
-    WebElement usernameField;
-
-    @FindBy(id = "password")
-    WebElement passwordField;
-
-    @FindBy(id = "loginButton")
-    WebElement loginButton;
-
-    @FindBy(xpath = "//textarea[@class='gLFyf']")
-    WebElement textBox;
-    public WebElement getTextBox(){
-        return textBox;
-    }
-
-    public LoginPage(WebDriver driver) {
-        this.driver = driver;
-        PageFactory.initElements(driver, this);
-    }
-
-    public void enterTextbox(String tb){
-        getTextBox().sendKeys(tb);
-        getTextBox().sendKeys(Keys.ENTER);
-    }
-
-
-    public void enterUsername(String username) {
-        usernameField.sendKeys(username);
-    }
-
-    public void enterPassword(String password) {
-        passwordField.sendKeys(password);
-    }
-
-    public void clickLoginButton() {
-        loginButton.click();
-    }
-}
