--- New BDD Project/src/test/java/utils/DecryptionUtil.java
@@ -1,9 +1,0 @@
-package utils;
-import jars.DataDecrypt;
-
-
-public class DecryptionUtil {
-    public static String decrypt(String encryptedPassword) throws Exception {
-        return DataDecrypt.decrypt(encryptedPassword);
-    }
-}
