--- New BDD Project/src/test/resources/features/RahulShettyLoginPracticePage.feature
@@ -1,7 +1,0 @@
-Feature: RahulShetty Login Functionality Practice
-
-  @run1
-  Scenario: Login page functionality practice
-    Given user launches the rahulshetty automation login page
-    When User enters the username and password
-    Then Validate login page details
