+++ New BDD Project/src/test/resources/features/loginPractice.feature
@@ -1,0 +1,7 @@
+Feature: RahulShetty Login Functionality Practice
+
+  @run
+  Scenario: Login page functionality practice
+    Given user launches the rahulshetty automation login page
+    When User enters the username and password
+    Then Validate login page details
